+++ packages/frontend/src/components/Settings/Advanced.tsx
@@ -1,17 +1,17 @@
 import React from 'react'
 
-import { SettingsStoreState } from '../../stores/settings'
+import type { SettingsStoreState } from '../../stores/settings'
 import Encryption from './Encryption'
 import { ExperimentalFeatures } from './ExperimentalFeatures'
 import ImapFolderHandling from './ImapFolderHandling'
 import SettingsHeading from './SettingsHeading'
 import SettingsSeparator from './SettingsSeparator'
 import useTranslationFunction from '../../hooks/useTranslationFunction'
-import Communication from './Communication'
 import EditAccountAndPasswordDialog from '../dialogs/EditAccountAndPasswordDialog'
 import useDialog from '../../hooks/dialog/useDialog'
 import SettingsButton from './SettingsButton'
 import { runtime } from '@deltachat-desktop/runtime-interface'
+import CoreSettingsSwitch from './CoreSettingsSwitch'
 
 type Props = {
   settingsStore: SettingsStoreState
@@ -26,7 +26,12 @@ export default function Advanced({ settingsStore }: Props) {
       <SettingsButton onClick={() => runtime.openLogFile()}>
         {tx('pref_view_log')}
       </SettingsButton>
-      <Communication settingsStore={settingsStore} />
+      <CoreSettingsSwitch
+        label={tx('enable_realtime')}
+        settingsKey='webxdc_realtime_enabled'
+        description={tx('enable_realtime_explain')}
+      />
+
       <SettingsSeparator />
 
       <SettingsHeading>{tx('pref_experimental_features')}</SettingsHeading>
@@ -47,7 +52,7 @@ export default function Advanced({ settingsStore }: Props) {
       >
         {tx('pref_password_and_account_settings')}
       </SettingsButton>
-      {settingsStore.settings.is_chatmail == '0' && (
+      {settingsStore.settings.is_chatmail === '0' && (
         <ImapFolderHandling settingsStore={settingsStore} />
       )}
     </>