+++ CHANGELOG.md
@@ -24,6 +24,7 @@
   - Stop saving txt_raw
   - Do not fail to send the message if some keys are missing
   - Synchronize contact name changes
+- hide 'show classic email' for chatmail, move down otherwise
 
 ### Fixed
 - tauri: improve security #4826